requirements.txt
Django==1.7.8
pinax-theme-bootstrap==5.6.0
django-user-accounts==1.0
metron==1.3.5
eventlog==0.8.0
django-jsonfield==0.9.13
django-timezones==0.2
pytz==2014.10
django-model-utils==2.2
django-taggit==0.12.2
django-reversion==1.8.5
django-markitup==2.2.2
markdown==2.6.2
django-sitetree==1.2.1
Pillow==2.6.1
easy-thumbnails==2.2
html5lib==0.999
django-multiselectfield==0.1.3

# @@@ Pin this once we get to stable release
-e git+git@github.com:toulibre/symposion.git@cdl#egg=symposion
